+++ src/SpeedyTransforms/show.jl
@@ -13,11 +13,16 @@ function prettymemory(b)
 end
 
 function Base.show(io::IO, S::SpectralTransform{NF, ArrayType}) where {NF, ArrayType}
-    (; lmax, mmax, Grid, nlat_half) = S
+    (; lmax, mmax, Grid, nlat_half, nlayers) = S
 
-    # add information about size of Legendre polynomials
-    s = S.recompute_legendre ? Base.summarysize(S.Λ) : Base.summarysize(S.Λs)
-    s_str = prettymemory(s)
+    # add information about size of Legendre polynomials and scratch memory
+    polysize_str = prettymemory(Base.summarysize(S.legendre_polynomials))
+    memorysize_str = prettymemory(
+                Base.summarysize(S.scratch_memory_north) +      # add all scratch_memories
+                Base.summarysize(S.scratch_memory_south) + 
+                Base.summarysize(S.scratch_memory_grid) + 
+                Base.summarysize(S.scratch_memory_spec)
+            )
 
     dealias = get_dealiasing(mmax, nlat_half)
     truncations = ["<linear", "linear", "quadratic", "cubic", ">cubic"]
@@ -28,5 +33,6 @@ function Base.show(io::IO, S::SpectralTransform{NF, ArrayType}) where {NF, Array
     println(io, "├ Spectral:   T$mmax, $(lmax+1)x$(mmax+1) LowerTriangularMatrix{Complex{$NF}}")
     println(io, "├ Grid:       $(RingGrids.get_nlat(Grid, nlat_half))-ring $Grid{$NF}")
     println(io, "├ Truncation: dealiasing = $dealiasing ($truncation)")
-      print(io, "└ Legendre:   recompute polynomials $(S.recompute_legendre) ($s_str)")
+    println(io, "├ Legendre:   Polynomials $polysize_str, shortcut: $(short_name(S.LegendreShortcut))")
+      print(io, "└ Memory:     for $nlayers layers ($memorysize_str)")
 end